+++ CHANGELOG.rst
@@ -4,6 +4,8 @@ Changelog
 1.15.0 (in development)
 -----------------------
 
+* Update :meth:`pname` method to support taxa names that don't have a delimiter such as ASV IDs (e.g. ``1ad289cd8f44e109fd95de0382c5b252``). Basically, the method will simply return the input name.
+
 1.14.0 (2022-06-24)
 -------------------